redis to redisson and cluster configs #77

+++ elements/badges/src/main/java/io/github/oasis/elements/badges/processors/StreakNBadgeProcessor.java
@@ -69,16 +69,16 @@ public class StreakNBadgeProcessor extends AbstractBadgeProcessor<StreakNBadgeRu
         if (rule.getCriteria().matches(event, rule, context)) {
             String member = ts + ONE_DELIM + event.getExternalId();
             sortedRange.add(member, ts);
-            long rank = sortedRange.getRank(member);
-            long start = Math.max(0, rank - rule.getMaxStreak());
+            int rank = sortedRange.getRank(member);
+            int start = Math.max(0, rank - rule.getMaxStreak());
             List<Record> tupleRange = sortedRange.getRangeByRankWithScores(start, rank + rule.getMaxStreak());
             return fold(tupleRange, event, rule, db);
         } else {
             String member = ts + ZERO_DELIM + event.getExternalId();
             sortedRange.add(member, ts);
             sortedRange.removeRangeByScore(0, ts - rule.getRetainTime());
-            long rank = sortedRange.getRank(member);
-            long start = Math.max(0, rank - rule.getMaxStreak());
+            int rank = sortedRange.getRank(member);
+            int start = Math.max(0, rank - rule.getMaxStreak());
             List<Record> tupleRange = sortedRange.getRangeByRankWithScores(start, rank + rule.getMaxStreak());
             return unfold(tupleRange, event, ts, rule, db);
         }
